requirements.txt
python-telegram-bot==13.15
aiogram==2.25.1
python-dotenv==1.0.0
SQLAlchemy==2.0.19 